requirements.txt
streamlit>=1.37.0
pandas>=2.2.2
numpy>=1.26.4
gspread>=6.1.4
google-auth>=2.32.0
requests>=2.32.3
beautifulsoup4>=4.12.3
openai>=1.44.0